fixing usecase invalid ip was being tried to be solved as hostname from internet
